nalgebra_sparse

Module csc

An implementation of the CSC sparse matrix format.

This is the module-level documentation. See CscMatrix for the main documentation of the CSC implementation.

Structs

CscCol

An immutable representation of a column in a CSC matrix.

CscColIter

Column iterator for CscMatrix.

CscColIterMut

Mutable column iterator for CscMatrix.

CscColMut

A mutable representation of a column in a CSC matrix.

CscMatrix

A CSC representation of a sparse matrix.

CscTripletIter

Iterator type for iterating over triplets in a CSC matrix.

CscTripletIterMut

Iterator type for mutably iterating over triplets in a CSC matrix.