nalgebra_sparse::ops::serial

Function spsolve_csc_lower_triangular

pub fn spsolve_csc_lower_triangular<'a, T: RealField>(
    l: Op<&CscMatrix<T>>,
    b: impl Into<DMatrixViewMut<'a, T>>,
) -> Result<(), OperationError>

Solve the lower triangular system op(L) X = B.

Only the lower triangular part of L is read, and the result is stored in B.

Errors

An error is returned if the system can not be solved due to the matrix being singular.

Panics

Panics if L is not square, or if L and B are not dimensionally compatible.